Version 1.3

+++ html/gitUpdate.php
@@ -4,12 +4,21 @@
 // echo "Please press back and refresh <br/>"
 //echo("<script>alert('Looking for the update .....')</script>");
 //exec("python /home/pi/auxilary.py");
-exec("git clone https://github.com/ApexPlus-Tech/NanoPiNeo_ps_and_at.git /var/www/updated",$output,$result);
-echo($result);
-//git cloning is successful
-if($result==0){
-	exec("cp -TRv /var/www/updated/html /var/www");
-	exec("rm -r --interactive=never /var/www/updated");
+exec("ping -c 4",$output,$result);
+if($result==0){//ping is successful
+	exec("git clone htt://github.com/ApexPlus-Tech/NanoPiNeo_ps_and_at.git /var/www/updated",$output,$result);
+	echo($result);
+	//git cloning is successful
+	if($result==0){
+		exec("cp -TRv /var/www/updated/html /var/www");
+		exec("rm -r --interactive=never /var/www/updated");
+		echo("Firmware update completed");
+	}
+	else{
+		echo("Server is down");
+	}
 }
-//echo("<script>window.location = 'index.html';</script>"); 
+else{
+	echo("Internet is down");
+}
 ?>
